LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/performance-timeline.https-expected.txt



PASS Test Performance Timeline API in Service Worker 
FAIL empty service worker fetch event included in performance timings assert_greater_than: Slow service worker request should measure increased delay. expected a number greater than 1152.1 but got 151.59999999999997
FAIL User Timing promise_test: Unhandled rejection with value: object "TypeError: performance.mark is not a function. (In 'performance.mark('startMark')', 'performance.mark' is undefined)"
FAIL Resource Timing promise_test: Unhandled rejection with value: object "TypeError: performance.getEntriesByType is not a function. (In 'performance.getEntriesByType('resource')', 'performance.getEntriesByType' is undefined)"
